EKF as of Thesis B
This is the first of the algorithm files in working condition to be uploaded, as of Thesis B. The rest will come in the coming week. It should be noted that the most significant change is that the a priori is initialised with the last estimate rather than the last measurement (mistake in the mathematics on my end).